PdfEncoding: Make the class final and non virtual
Also made PdfEncoding::ExportToFont() private/internal

+++ src/podofo/main/PdfFontManager.h
@@ -161,7 +161,7 @@ private:
 
         std::string Name;               ///< Name of the font or pattern
         PdfStandard14FontType StdType;
-        size_t EncodingId;
+        unsigned EncodingId;
         bool HasFontStyle;
         PdfFontStyle Style;
     };
